Add files via upload
The usual method for doing this is to contrast the model's predictions with real gold prices. The model must be used to produce predictions on new data in a real-world scenario as the last stage before deployment. The neural network is one of the most used machine learning techniques for predicting the price of gold.
Forecasting rise and decline in the daily gold rates, can help investors to decide when to buy (or sell) the commodity